@@ -1,24 +1,21 @@
 package de.elnarion.asciidoc.test;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 
 import org.asciidoctor.Asciidoctor;
 import org.asciidoctor.Asciidoctor.Factory;
-import org.asciidoctor.Attributes;
 import org.asciidoctor.Options;
 import org.asciidoctor.SafeMode;
 import org.asciidoctor.extension.RubyExtensionRegistry;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 /**
  * The Class TestExtension.
  */
-public class TestExtension {
+class TestExtension {
 
 	/**
 	 * Test extension.
@@ -26,22 +23,13 @@ public class TestExtension {
 	 * @throws IOException
 	 */
 	@Test
-	public void testExtension() throws IOException {
+	void testExtension() throws IOException {
 		Asciidoctor asciidoctor = Factory.create();
 		RubyExtensionRegistry rubyReg = asciidoctor.rubyExtensionRegistry();
 		rubyReg.loadClass(
 				TestExtension.class.getClassLoader().getResourceAsStream("elnarion/multi-include-processor/extension.rb"));
 		rubyReg.includeProcessor("MultiIncludeProcessor");
-		Attributes asciidocAttributes = new Attributes();
-		asciidocAttributes.setAllowUriRead(true);
-		asciidocAttributes.setBackend("docbook");
-		Map<String, Object> attributes = new HashMap<String, Object>();
-		attributes.put("backend", "docbook");
-		attributes.put("allow-uri-read", "true");
-		Options asciidocOptions = new Options();
-		asciidocOptions.setSafe(SafeMode.UNSAFE);
-		asciidocOptions.setInPlace(true);
-		asciidocOptions.setToFile(false);
+		Options asciidocOptions =Options.builder().safe(SafeMode.UNSAFE).inPlace(true).toFile(false).build();
 		String testResult = asciidoctor.convertFile(new File("./src/test/resources/multitest/test.adoc"), asciidocOptions);
 		assertTrue(testResult.contains("multixcontent2"));
 
